Remove unnecessary prerequisites (#9)
* Handle the scenario where the release version already exists and so there's no need to make one

+++ scripts/release.sh
@@ -48,15 +48,21 @@ res=`curl --user "$GH_USER:$GH_PATH" -X POST https://api.github.com/repos/${GH_U
 }"`
 echo Create release result: ${res}
 rel_id=`echo ${res} | python -c 'import json,sys;print(json.load(sys.stdin)["id"])'`
-file_name=${NAME}-${VERSION}.tar.gz
+if [ -z "$rel_id" ]
+then
+  echo "No release required as it already exists"
+else
+  file_name=${NAME}-${VERSION}.tar.gz
+  curl --user "$GH_USER:$GH_PATH" -X POST https://uploads.github.com/repos/${GH_USER}/${GH_REPO}/releases/${rel_id}/assets?name=${file_name}\
+    --header 'Content-Type: text/javascript ' --upload-file ${ASSETS_PATH}/${file_name}
 
-curl --user "$GH_USER:$GH_PATH" -X POST https://uploads.github.com/repos/${GH_USER}/${GH_REPO}/releases/${rel_id}/assets?name=${file_name}\
- --header 'Content-Type: text/javascript ' --upload-file ${ASSETS_PATH}/${file_name}
+  source ./scripts/package_homebrew.sh
 
-source ./scripts/package_homebrew.sh
+  rm ${ASSETS_PATH}/${file_name}
+fi
 
 # clean up
-rm ${ASSETS_PATH}/${file_name}
+
 rm hash.txt
 rm version.txt
 rm package.txt
